Search fix due to wrong type hinting
In the `Magentix_Solr_Model_Search::loadQuery()` method the `$response` returned by `$this->search()` is an instance of `Apache_Solr_Response` but then the response is taken from the `response` parsed data (by doing `$response->response`) which is an instance of `stdClass` and not `Apache_Solr_Response`. So the type hinting on `Magentix_Solr_Model_Search::setResponse()` was wrong.